+++ noisetools/hawc2.py
@@ -57,13 +57,18 @@ def read_hawc2_res(model_path: str | os.PathLike,
     fname = os.path.join(model_path, output_filename)
     # Read the sel file
     sel_file = pd.read_fwf(fname + '.sel', skiprows=12, header=None, skipfooter=1, index_col=0,
-                           widths=[12, 32, 10, 120], )
+                           colspecs=[(0, 13), (13, 44), (44, 55), (55, None)]
+                           )
+
     # Read the data file and shift the column numbers to correspond to the channel numbers
     dat_file = pd.read_csv(fname + '.dat', delimiter='\\s+', header=None, index_col=0, )
-    if numbered_columns:
-        dat_file.columns = dat_file.columns + 1
-    else:
-        dat_file.columns = list(sel_file.loc[2:, 1])
+    dat_file.columns = dat_file.columns + 1
+
+    # Read the column names from the sel file if numbered columns is turned off.
+    if not numbered_columns:
+        dat_file.columns = sel_file.loc[dat_file.columns, 1].values
+
+    # Round the index to remove machine error.
     dat_file.index = np.round(dat_file.index, 9)
 
     return sel_file, dat_file